refactor: Entity 분리
