+++ src/test/java/chess/domain/ChessGameTest.java
@@ -8,7 +8,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 import chess.domain.command.MoveCommand;
 import chess.domain.game.ChessGame;
 import chess.domain.game.GameResult;
-import chess.domain.game.Room;
+import chess.entity.Room;
 import chess.domain.piece.ChessmenInitializer;
 import chess.domain.piece.Color;
 import chess.domain.piece.King;
